Add API for talking to the database
Start by providing a simple to use GetSetting() function that allows us
to fetch a user-specific value for a setting, or a defualt value if a
user-specific value is not available.

No test for a valid setting name is currently performed (you'll just get
an empty string back).

See also (from within the `backend` directory):

```
go doc internal/database.GetSetting
```