Turn of SQL linter
The SQLFluff linter knows a number of different SQL dialects, and SQLite
is supposed to be one of these.  However, even when configuring the
linter to use the SQLite dialect, the linter does not understand the
`CREATE TABLE` statements, or the `PRAGMA` statements in the SQL files
that we have.  I'm therefore suggesting turning this linter off.

I will, however, possibly sort out the indentation of the SQL files,
which SQLFluff has suggestions about.  See,
https://docs.sqlfluff.com/en/stable/indentation.html

There are also obvious quoting errors that need to be fixed.