Refactoring
Rename both the `/report` and the `/spent_time` endpoints
`/time_entries`, mimicking the Redmine REST API.  This is done to reduce
confusion.

Allow the GetTimeEntries() function in our Redmine API to get optional
parameters for filtering, which reduses the number of neede functions
for the various use cases.

Pass the fetched time entries from Redmien directly on te the frontend.
This is done to reduce the amount of processing that the backend is
doing.  Additionally, this gives the frontend all the data it evern
needs with regards to the time entries fetched from Redmine.