Unpack the essential bits of the request and proxy depending on what we find
Only the `id` and `hours` part of the `time_entry` structure is needed
to decide how to forward the request to Redmine.  Depending on these
values, we create a new time entry, update an existing one, or delete an
existing time entry.

This commit is the merging of two older commits.  The renaming of
`CreateTimeEntry()` --> `ProcessTimeEntry()` is essentially a no-op in
this commit as that function is immediately deleted.  The fact that the
`CreateTimeEntry()` goes through a step of being renamed first is due to
a mishap during a rebase.