Add custom name to the `issueActivity` struct
This means that the response from the `/api/recent_issues` endpoint will
contain empty strings in this field.  However, the `/api/favorites`
endpoint will be able to use exactly the same data model and _will_ fill
out this field.

This commit also fixes an awkward creation of a slice in the `favorite`
package.