Remove use of external Docker network
Corresponds to the live changes made on the 31st of October where a
Docker-related networking issue forced us to move the service into an
LXC container.  The use of an external Docker network was a hack to cope
with an earlier Docker network issue.  That issue was resolved when
setting up the service in LXC.