fix: Figure out the Git data in the Dockerfile
Writes the correct environment variables to a file that gets sourced by
the `entrypoint.sh` script.  If the variables are already set in the
environment, then those given values will be used instead of running
`git` commands.

The `git` utility is installed temporarily in the Dockerfile to get the
data.  The `.git` directory is likewise copied into the image to get the
data, but is removed before the image is finished building.